requirements.txt
# FastAPI framework and server
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0

# Database
sqlalchemy==2.0.21
alembic==1.12.0
psycopg2-binary==2.9.7
supabase==1.0.4

# AI providers
openai==1.3.3
anthropic==0.5.0

# Redis caching (optional)
redis==4.6.0

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-bdd==7.0.0
httpx==0.24.1  # Compatible with supabase

# Code quality
black==23.9.1
isort==5.12.0
mypy==1.5.1
flake8==6.1.0

# Utils
python-dotenv==1.0.0
pydantic-settings==2.0.3
tenacity==8.2.3
